refactor: mv generation config to new file
