+++ discojs/src/models/implementations/gpt/config.ts
@@ -46,7 +46,7 @@ export const DefaultGPTConfig: Required<GPTConfig> = {
   nLayer: 3,
   nHead: 3,
   nEmbd: 48,
-  seed: Math.random(),
+  seed: Math.floor(Math.random() * Number.MAX_SAFE_INTEGER),
 };
 
 export type ModelSize = {
@@ -73,24 +73,3 @@ export function getModelSizes(modelType: GPTModelType): Required<ModelSize> {
       return { nLayer: 3, nHead: 3, nEmbd: 48 };
   }
 }
-
-export interface GenerationConfig {
-  // take random token weighted by its probability
-  // If false, predict the token with the highest probability.
-  doSample: boolean;
-  // the generation temperature (higher means more randomness).
-  // Set to 0 for greedy decoding.
-  temperature: number;
-  // only consider the topk most likely tokens for sampling.
-  // used if doSample is true.
-  topk: number;
-  // random seed for sampling.
-  seed: number;
-}
-
-export const DefaultGenerationConfig: Required<GenerationConfig> = {
-  temperature: 1.0,
-  doSample: false,
-  seed: Math.random(),
-  topk: 50,
-};
