+++ discojs/src/models/index.ts
@@ -2,6 +2,8 @@ export { Model } from "./model.js";
 export type { BatchLogs, ValidationMetrics } from "./logs.js";
 export { EpochLogs } from "./logs.js";
 export { Tokenizer } from "./tokenizer.js";
+export { DefaultGenerationConfig } from "./generation.js";
+export type { GenerationConfig } from "./generation.js";
 
 export type {
   GPTConfig,
