+++ discojs/src/models/onnx.ts
@@ -5,9 +5,9 @@ import { List } from "immutable";
 import type { WeightsContainer } from "#weights/index";
 import type { Batched } from "#dataset/index";
 import type { DataFormat } from "#types/index";
-import type { GenerationConfig as TFJSGenerationConfig } from "#models/implementations/gpt/config";
 import { Model } from "#models/model";
-import { DefaultGenerationConfig } from "#models/implementations/gpt/config";
+import { DefaultGenerationConfig } from "#models/generation";
+import type { GenerationConfig } from "#models/generation";
 
 export class ONNXModel extends Model<"text"> {
   readonly datatype = "text" as const;
@@ -30,7 +30,7 @@ export class ONNXModel extends Model<"text"> {
 
   override async predict(
     batch: Batched<DataFormat.ModelEncoded["text"][0]>,
-    options?: Partial<TFJSGenerationConfig>,
+    options?: Partial<GenerationConfig>,
   ): Promise<Batched<DataFormat.ModelEncoded["text"][1]>> {
     const config = Object.assign({}, DefaultGenerationConfig, options);
 
@@ -43,7 +43,7 @@ export class ONNXModel extends Model<"text"> {
 
   async #predictSingle(
     tokens: DataFormat.ModelEncoded["text"][0],
-    config: TFJSGenerationConfig,
+    config: GenerationConfig,
   ): Promise<DataFormat.ModelEncoded["text"][1]> {
     const contextLength =
       (this.model.config as { max_position_embeddings?: number })
